test: ensure text buffer get lines bytes works

+++ lib/buffers/gap/buffer.v
@@ -62,8 +62,8 @@ fn (gb Buffer) gap_size() u64 {
 	return gb.cend - gb.ccur
 }
 
-pub fn (gb Buffer) logical_len() int {
-	return gb.buf.len - int(gb.gap_size())
+pub fn (gb Buffer) logical_len() u64 {
+	return u64(gb.buf.len - int(gb.gap_size()))
 }
 
 pub fn (gb Buffer) get(pos u64) ?u8 {
